refactor(mtp): split implementation into modules
