+++ device/esp_tinyusb/Kconfig
@@ -169,7 +169,6 @@ menu "TinyUSB Stack"
         config TINYUSB_MTP_EP_BUFSIZE
             depends on TINYUSB_MTP_ENABLED
             int "MTP endpoint buffer size"
-            default 4096 if IDF_TARGET_ESP32P4 || IDF_TARGET_ESP32S31
             default 4096
             range 4096 8192 if IDF_TARGET_ESP32S2 || IDF_TARGET_ESP32S3 || IDF_TARGET_ESP32H4
             range 4096 32768 if IDF_TARGET_ESP32P4 || IDF_TARGET_ESP32S31
