+++ device/esp_tinyusb/include_private/mtp/tinyusb_mtp_operations.h
@@ -1,0 +1,18 @@
+/*
+ * SPDX-FileCopyrightText: 2026 Espressif Systems (Shanghai) CO LTD
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+#pragma once
+
+#include "tusb.h"
+
+int32_t mtp_op_get_device_info(tud_mtp_cb_data_t *cb_data);
+int32_t mtp_op_open_close_session(tud_mtp_cb_data_t *cb_data);
+int32_t mtp_op_get_storage_ids(tud_mtp_cb_data_t *cb_data);
+int32_t mtp_op_get_storage_info(tud_mtp_cb_data_t *cb_data);
+int32_t mtp_op_get_num_objects(tud_mtp_cb_data_t *cb_data);
+int32_t mtp_op_get_object_handles(tud_mtp_cb_data_t *cb_data);
+int32_t mtp_op_get_object_info(tud_mtp_cb_data_t *cb_data);
+int32_t mtp_op_delete_object(tud_mtp_cb_data_t *cb_data);
+int32_t mtp_op_move_object(tud_mtp_cb_data_t *cb_data);
