+++ device/esp_tinyusb/include_private/mtp/tinyusb_mtp_write.h
@@ -1,0 +1,15 @@
+/*
+ * SPDX-FileCopyrightText: 2026 Espressif Systems (Shanghai) CO LTD
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+#pragma once
+
+#include "tusb.h"
+
+int32_t mtp_op_begin_edit_object(tud_mtp_cb_data_t *cb_data);
+int32_t mtp_op_send_partial_object(tud_mtp_cb_data_t *cb_data);
+int32_t mtp_op_truncate_object(tud_mtp_cb_data_t *cb_data);
+int32_t mtp_op_end_edit_object(tud_mtp_cb_data_t *cb_data);
+int32_t mtp_op_send_object_info(tud_mtp_cb_data_t *cb_data);
+int32_t mtp_op_send_object(tud_mtp_cb_data_t *cb_data);
