+++ device/esp_tinyusb/test_apps/mtp_storage/main/test_mtp_storage.c
@@ -6,7 +6,9 @@
 
 #include "soc/soc_caps.h"
 
+#include <dirent.h>
 #include <errno.h>
+#include <inttypes.h>
 #include <stdbool.h>
 #include <stdio.h>
 #include <string.h>
@@ -19,7 +21,9 @@
 #include "tinyusb.h"
 #include "tinyusb_default_config.h"
 #include "tinyusb_mtp.h"
-#include "tinyusb_mtp_impl.h"
+#include "mtp/tinyusb_mtp_context.h"
+#include "mtp/tinyusb_mtp_object_store.h"
+#include "mtp/tinyusb_mtp_types.h"
 #include "tinyusb_mtp_test.h"
 #include "unity.h"
 #include "wear_levelling.h"
@@ -49,10 +53,31 @@ static bool s_mtp_storage_warmed_up;
 
 void test_mtp_storage_warm_up(void);
 
+typedef struct {
+    TaskHandle_t owner;
+    TaskHandle_t task;
+    volatile bool stop;
+} test_mtp_callback_race_t;
+
+static void test_mtp_callback_race_task(void *arg)
+{
+    test_mtp_callback_race_t *race = arg;
+    while (!race->stop) {
+        (void)tud_mtp_request_cancel_cb(NULL);
+        vTaskDelay(1);
+    }
+    xTaskNotifyGive(race->owner);
+    vTaskSuspend(NULL);
+}
+
 static const char *const s_test_artifacts[] = {
     "hello.txt",
     "empty.txt",
     "empty_replace.txt",
+    "empty_storage_zero.txt",
+    "active_pending.txt",
+    "cancel_first.txt",
+    "cancel_second.txt",
     "direct_edit.txt",
     "cross_edit.txt",
     "4G-usb.md",
@@ -117,6 +142,28 @@ static void test_mtp_clean_storage(test_mtp_storage_t *storage)
         test_mtp_build_path(path, sizeof(path), storage->base_path, s_test_artifacts[i]);
         test_mtp_remove_path_if_exists(path);
     }
+
+    while (true) {
+        DIR *dir = opendir(storage->base_path);
+        TEST_ASSERT_NOT_NULL_MESSAGE(dir, "Failed to open MTP test storage for cleanup");
+        char stale_name[256] = { 0 };
+        struct dirent *entry;
+        while ((entry = readdir(dir)) != NULL) {
+            if (strncmp(entry->d_name, "ci_perf_", 8) == 0 || strncmp(entry->d_name, "ci_concurrent_", 14) == 0 || strncmp(entry->d_name, "ci_many_", 8) == 0 ||
+                    strncmp(entry->d_name, "ci_cache_", 9) == 0) {
+                int len = snprintf(stale_name, sizeof(stale_name), "%s", entry->d_name);
+                TEST_ASSERT_TRUE_MESSAGE(len > 0 && len < (int)sizeof(stale_name), "MTP stale workspace name was truncated");
+                break;
+            }
+        }
+        TEST_ASSERT_EQUAL(0, closedir(dir));
+        if (stale_name[0] == '\0') {
+            break;
+        }
+        char path[128];
+        test_mtp_build_path(path, sizeof(path), storage->base_path, stale_name);
+        TEST_ASSERT_EQUAL_MESSAGE(ESP_OK, mtp_recursive_delete_path(path), "Failed to remove stale MTP host test workspace");
+    }
 }
 
 static void test_mtp_clean_all_storages(void)
@@ -136,6 +183,19 @@ static void test_mtp_clean_storage_path(test_mtp_storage_t *storage, const char 
 static void test_mtp_write_file(const char *path, const char *content)
 {
     FILE *file = fopen(path, "wb");
+    if (file == NULL) {
+        ESP_LOGE(TAG, "failed to open MTP test file %s for write: %s", path, strerror(errno));
+        for (size_t i = 0; i < TEST_MTP_STORAGE_COUNT; i++) {
+            size_t base_len = strlen(s_test_storages[i].base_path);
+            if (strncmp(path, s_test_storages[i].base_path, base_len) == 0 && path[base_len] == '/') {
+                uint64_t total = 0;
+                uint64_t free_bytes = 0;
+                esp_err_t ret = esp_vfs_fat_info(s_test_storages[i].base_path, &total, &free_bytes);
+                ESP_LOGE(TAG, "storage state: path=%s info=%s total=%" PRIu64 " free=%" PRIu64, s_test_storages[i].base_path, esp_err_to_name(ret), total, free_bytes);
+                break;
+            }
+        }
+    }
     TEST_ASSERT_NOT_NULL_MESSAGE(file, "Failed to open MTP test file for write");
     TEST_ASSERT_EQUAL(strlen(content), fwrite(content, 1, strlen(content), file));
     TEST_ASSERT_EQUAL(0, fclose(file));
@@ -240,12 +300,27 @@ static size_t test_mtp_cached_object_count(tinyusb_mtp_storage_handle_t storage)
 {
     size_t count = 0;
     mtp_lock();
-    for (size_t i = 0; i < CONFIG_TINYUSB_MTP_MAX_OBJECTS; i++) {
-        if (s_mtp.mux_protected.objects[i].used && s_mtp.mux_protected.objects[i].storage == storage) {
-            count++;
-        }
-    }
+    count = mtp_object_count_for_storage_locked(storage);
     mtp_unlock();
+    return count;
+}
+
+static uint32_t test_mtp_get_object_count(uint16_t operation, uint32_t storage_id, uint32_t object_format, uint32_t parent)
+{
+    uint8_t response[sizeof(uint32_t) + CONFIG_TINYUSB_MTP_MAX_OBJECTS * sizeof(uint32_t)] = { 0 };
+    tinyusb_mtp_test_result_t result;
+    TEST_ASSERT_EQUAL(ESP_OK, tinyusb_mtp_test_execute(&(tinyusb_mtp_test_transaction_t) {
+        .operation = operation,
+        .phase = MTP_PHASE_COMMAND,
+        .param_count = 3,
+        .params = { storage_id, object_format, parent },
+        .complete_data = operation == MTP_OP_GET_OBJECT_HANDLES,
+    }, response, sizeof(response), &result));
+    TEST_ASSERT_EQUAL_INT32(MTP_RESP_OK, result.response_code);
+    TEST_ASSERT_GREATER_OR_EQUAL_UINT32(sizeof(uint32_t), result.data_len);
+
+    uint32_t count = 0;
+    memcpy(&count, response, sizeof(count));
     return count;
 }
 
@@ -321,11 +396,29 @@ TEST_CASE("MTP: register mounted FATFS and delete object through backend", "[mtp
     TEST_ASSERT_EQUAL(ESP_OK, tinyusb_mtp_uninstall_driver());
 }
 
+TEST_CASE("MTP: uninstall serializes with callbacks", "[mtp][storage][ci]")
+{
+    test_mtp_callback_race_t race = {
+        .owner = xTaskGetCurrentTaskHandle(),
+    };
+    TEST_ASSERT_EQUAL(pdPASS, xTaskCreate(test_mtp_callback_race_task, "mtp_cb_race", 3072, &race, 5, &race.task));
+    for (size_t i = 0; i < 20; i++) {
+        TEST_ASSERT_EQUAL(ESP_OK, tinyusb_mtp_install_driver(NULL));
+        taskYIELD();
+        TEST_ASSERT_EQUAL(ESP_OK, tinyusb_mtp_uninstall_driver());
+    }
+    race.stop = true;
+    TEST_ASSERT_EQUAL(1, ulTaskNotifyTake(pdTRUE, pdMS_TO_TICKS(1000)));
+    vTaskDelete(race.task);
+    vTaskDelay(1);
+}
+
 TEST_CASE("MTP: zero-size SendObjectInfo completes without SendObject", "[mtp][storage][ci]")
 {
     test_mtp_storage_t *flash = &s_test_storages[0];
     test_mtp_clean_storage_path(flash, "empty.txt");
     test_mtp_clean_storage_path(flash, "empty_replace.txt");
+    test_mtp_clean_storage_path(flash, "empty_storage_zero.txt");
 
     TEST_ASSERT_EQUAL(ESP_OK, tinyusb_mtp_install_driver(NULL));
     tinyusb_mtp_storage_handle_t storage = NULL;
@@ -337,9 +430,17 @@ TEST_CASE("MTP: zero-size SendObjectInfo completes without SendObject", "[mtp][s
     TEST_ASSERT_NOT_NULL(storage);
 
     uint32_t empty_handle = 0;
-    TEST_ASSERT_EQUAL(ESP_OK, tinyusb_mtp_test_send_zero_size_object_info(storage, MTP_ROOT_PARENT, "empty.txt", &empty_handle));
+    TEST_ASSERT_EQUAL(ESP_OK, tinyusb_mtp_test_send_object_info(storage, TINYUSB_MTP_TEST_STORAGE_ID_REGISTERED, TINYUSB_MTP_TEST_STORAGE_ID_REGISTERED,
+                                                                MTP_ROOT_PARENT, "empty.txt", 0, &empty_handle));
     TEST_ASSERT_NOT_EQUAL(0, empty_handle);
-    TEST_ASSERT_EQUAL(ESP_OK, tinyusb_mtp_test_send_zero_size_object());
+    uint8_t response_params[3 * sizeof(uint32_t)] = { 0 };
+    tinyusb_mtp_test_result_t send_result;
+    TEST_ASSERT_EQUAL(ESP_OK, tinyusb_mtp_test_execute(&(tinyusb_mtp_test_transaction_t) {
+        .operation = MTP_OP_SEND_OBJECT,
+        .phase = MTP_PHASE_DATA,
+        .complete_data = true,
+    }, response_params, sizeof(response_params), &send_result));
+    TEST_ASSERT_EQUAL_INT32(MTP_RESP_OK, send_result.response_code);
 
     uint32_t parent_handle = 0;
     TEST_ASSERT_EQUAL(ESP_OK, tinyusb_mtp_test_get_parent_handle(empty_handle, &parent_handle));
@@ -352,13 +453,166 @@ TEST_CASE("MTP: zero-size SendObjectInfo completes without SendObject", "[mtp][s
     const char *replace_path = TEST_MTP_FLASH_BASE_PATH "/empty_replace.txt";
     test_mtp_write_file(replace_path, "replace me");
     uint32_t replace_handle = 0;
-    TEST_ASSERT_EQUAL(ESP_OK, tinyusb_mtp_test_send_zero_size_object_info(storage, MTP_ROOT_PARENT, "empty_replace.txt", &replace_handle));
+    TEST_ASSERT_EQUAL(ESP_OK, tinyusb_mtp_test_send_object_info(storage, TINYUSB_MTP_TEST_STORAGE_ID_REGISTERED, TINYUSB_MTP_TEST_STORAGE_ID_REGISTERED,
+                                                                MTP_ROOT_PARENT, "empty_replace.txt", 0, &replace_handle));
     TEST_ASSERT_NOT_EQUAL(0, replace_handle);
     TEST_ASSERT_EQUAL(0, stat(replace_path, &st));
     TEST_ASSERT_EQUAL(0, (int)st.st_size);
 
+    uint32_t zero_storage_handle = 0;
+    TEST_ASSERT_EQUAL(ESP_OK, tinyusb_mtp_test_send_object_info(storage, MTP_STORAGE_ID_UNSPECIFIED, MTP_STORAGE_ID_UNSPECIFIED, MTP_PARENT_ROOT,
+                                                                "empty_storage_zero.txt", 0, &zero_storage_handle));
+    TEST_ASSERT_NOT_EQUAL(0, zero_storage_handle);
+    TEST_ASSERT_EQUAL(0, stat(TEST_MTP_FLASH_BASE_PATH "/empty_storage_zero.txt", &st));
+
     TEST_ASSERT_EQUAL(ESP_OK, tinyusb_mtp_unregister_storage(storage));
     TEST_ASSERT_EQUAL(ESP_OK, tinyusb_mtp_uninstall_driver());
+}
+
+TEST_CASE("MTP: active USB transfer blocks storage unregister", "[mtp][storage][ci]")
+{
+    test_mtp_storage_t *flash = &s_test_storages[0];
+    test_mtp_clean_storage_path(flash, "active_pending.txt");
+    TEST_ASSERT_EQUAL(ESP_OK, tinyusb_mtp_install_driver(NULL));
+    tinyusb_mtp_storage_handle_t storage = NULL;
+    TEST_ASSERT_EQUAL(ESP_OK, tinyusb_mtp_register_storage(&(tinyusb_mtp_storage_config_t) {
+        .base_path = flash->base_path,
+        .display_name = flash->display_name,
+        .removable = flash->removable,
+    }, &storage));
+
+    uint32_t object_handle = 0;
+    TEST_ASSERT_EQUAL(ESP_OK, tinyusb_mtp_test_send_object_info(storage, TINYUSB_MTP_TEST_STORAGE_ID_REGISTERED, TINYUSB_MTP_TEST_STORAGE_ID_REGISTERED,
+                                                                MTP_ROOT_PARENT, "active_pending.txt", 1, &object_handle));
+    TEST_ASSERT_NOT_EQUAL(0, object_handle);
+    TEST_ASSERT_EQUAL(ESP_ERR_INVALID_STATE, tinyusb_mtp_unregister_storage(storage));
+    TEST_ASSERT_TRUE(tud_mtp_request_cancel_cb(NULL));
+    TEST_ASSERT_EQUAL(ESP_OK, tinyusb_mtp_unregister_storage(storage));
+    TEST_ASSERT_EQUAL(ESP_OK, tinyusb_mtp_uninstall_driver());
+    test_mtp_clean_storage_path(flash, "active_pending.txt");
+}
+
+TEST_CASE("MTP: cancelled read allows the next object read", "[mtp][storage][ci]")
+{
+    test_mtp_storage_t *flash = &s_test_storages[0];
+    char first_path[96];
+    char second_path[96];
+    test_mtp_build_path(first_path, sizeof(first_path), flash->base_path, "cancel_first.txt");
+    test_mtp_build_path(second_path, sizeof(second_path), flash->base_path, "cancel_second.txt");
+    test_mtp_write_file(first_path, "first object content must span the test packet so cancellation occurs before transfer completion");
+    test_mtp_write_file(second_path, "second object");
+
+    TEST_ASSERT_EQUAL(ESP_OK, tinyusb_mtp_install_driver(NULL));
+    tinyusb_mtp_storage_handle_t storage = NULL;
+    TEST_ASSERT_EQUAL(ESP_OK, tinyusb_mtp_register_storage(&(tinyusb_mtp_storage_config_t) {
+        .base_path = flash->base_path,
+        .display_name = flash->display_name,
+        .removable = flash->removable,
+    }, &storage));
+    uint32_t first_handle = 0;
+    uint32_t second_handle = 0;
+    TEST_ASSERT_EQUAL(ESP_OK, tinyusb_mtp_test_find_object(storage, first_path, &first_handle));
+    TEST_ASSERT_EQUAL(ESP_OK, tinyusb_mtp_test_find_object(storage, second_path, &second_handle));
+    TEST_ASSERT_EQUAL(ESP_OK, tinyusb_mtp_test_cancel_read_and_restart(first_handle, second_handle));
+    TEST_ASSERT_EQUAL(ESP_OK, tinyusb_mtp_unregister_storage(storage));
+    TEST_ASSERT_EQUAL(ESP_OK, tinyusb_mtp_uninstall_driver());
+    unlink(first_path);
+    unlink(second_path);
+}
+
+TEST_CASE("MTP: all-object property list traverses all depths", "[mtp][storage][ci]")
+{
+    test_mtp_storage_t *flash = &s_test_storages[0];
+    char dir_path[128];
+    char child_path[160];
+    test_mtp_build_path(dir_path, sizeof(dir_path), flash->base_path, "ci_prop_tree");
+    test_mtp_build_path(child_path, sizeof(child_path), dir_path, "child.txt");
+    struct stat st;
+    if (stat(dir_path, &st) == 0) {
+        TEST_ASSERT_EQUAL(ESP_OK, mtp_recursive_delete_path(dir_path));
+    }
+    test_mtp_make_dir_if_missing(dir_path);
+    test_mtp_write_file(child_path, "property");
+
+    TEST_ASSERT_EQUAL(ESP_OK, tinyusb_mtp_install_driver(NULL));
+    tinyusb_mtp_storage_handle_t storage = NULL;
+    TEST_ASSERT_EQUAL(ESP_OK, tinyusb_mtp_register_storage(&(tinyusb_mtp_storage_config_t) {
+        .base_path = flash->base_path,
+        .display_name = flash->display_name,
+        .removable = flash->removable,
+    }, &storage));
+
+    uint32_t counts[2] = { 0 };
+    const uint32_t depths[] = { 0, MTP_DEPTH_ALL };
+    for (size_t i = 0; i < 2; i++) {
+        uint8_t response[1024] = { 0 };
+        tinyusb_mtp_test_result_t result;
+        TEST_ASSERT_EQUAL(ESP_OK, tinyusb_mtp_test_execute(&(tinyusb_mtp_test_transaction_t) {
+            .operation = MTP_OP_GET_OBJECT_PROP_LIST,
+            .phase = MTP_PHASE_COMMAND,
+            .param_count = 5,
+            .params = { MTP_OBJECT_HANDLE_ALL, 0, MTP_OBJ_PROP_OBJECT_FILE_NAME, 0, depths[i] },
+            .complete_data = true,
+        }, response, sizeof(response), &result));
+        TEST_ASSERT_EQUAL_INT32(MTP_RESP_OK, result.response_code);
+        TEST_ASSERT_GREATER_OR_EQUAL_UINT32(sizeof(uint32_t), result.data_len);
+        memcpy(&counts[i], response, sizeof(counts[i]));
+    }
+    TEST_ASSERT_GREATER_OR_EQUAL_UINT32(2, counts[0]);
+    TEST_ASSERT_EQUAL_UINT32(counts[1], counts[0]);
+
+    TEST_ASSERT_EQUAL(ESP_OK, tinyusb_mtp_unregister_storage(storage));
+    TEST_ASSERT_EQUAL(ESP_OK, tinyusb_mtp_uninstall_driver());
+    TEST_ASSERT_EQUAL(ESP_OK, mtp_recursive_delete_path(dir_path));
+}
+
+TEST_CASE("MTP: all-object handle queries traverse all depths", "[mtp][storage][ci]")
+{
+    test_mtp_storage_t *flash = &s_test_storages[0];
+    char tree_path[128];
+    char root_file[160];
+    char dir_path[160];
+    char nested_file[192];
+    char subdir_path[192];
+    char deep_file[224];
+    test_mtp_build_path(tree_path, sizeof(tree_path), flash->base_path, "ci_handle_tree");
+    test_mtp_build_path(root_file, sizeof(root_file), tree_path, "root.txt");
+    test_mtp_build_path(dir_path, sizeof(dir_path), tree_path, "folder");
+    test_mtp_build_path(nested_file, sizeof(nested_file), dir_path, "nested.mp3");
+    test_mtp_build_path(subdir_path, sizeof(subdir_path), dir_path, "subfolder");
+    test_mtp_build_path(deep_file, sizeof(deep_file), subdir_path, "deep.txt");
+
+    struct stat st;
+    if (stat(tree_path, &st) == 0) {
+        TEST_ASSERT_EQUAL(ESP_OK, mtp_recursive_delete_path(tree_path));
+    }
+
+    TEST_ASSERT_EQUAL(ESP_OK, tinyusb_mtp_install_driver(NULL));
+    tinyusb_mtp_storage_handle_t storage = NULL;
+    TEST_ASSERT_EQUAL(ESP_OK, tinyusb_mtp_register_storage(&(tinyusb_mtp_storage_config_t) {
+        .base_path = flash->base_path,
+        .display_name = "Handle tree",
+    }, &storage));
+
+    uint32_t root_count = test_mtp_get_object_count(MTP_OP_GET_OBJECT_HANDLES, storage->storage_id, 0, MTP_OBJECT_HANDLE_INVALID);
+    uint32_t all_count = test_mtp_get_object_count(MTP_OP_GET_OBJECT_HANDLES, storage->storage_id, 0, MTP_OBJECT_HANDLE_ALL);
+    uint32_t mp3_count = test_mtp_get_object_count(MTP_OP_GET_OBJECT_HANDLES, storage->storage_id, MTP_OBJ_FORMAT_MP3, MTP_OBJECT_HANDLE_ALL);
+
+    test_mtp_make_dir_if_missing(tree_path);
+    test_mtp_write_file(root_file, "root");
+    test_mtp_make_dir_if_missing(dir_path);
+    test_mtp_write_file(nested_file, "audio");
+    test_mtp_make_dir_if_missing(subdir_path);
+    test_mtp_write_file(deep_file, "deep");
+
+    TEST_ASSERT_EQUAL_UINT32(root_count + 1, test_mtp_get_object_count(MTP_OP_GET_OBJECT_HANDLES, storage->storage_id, 0, MTP_OBJECT_HANDLE_INVALID));
+    TEST_ASSERT_EQUAL_UINT32(all_count + 6, test_mtp_get_object_count(MTP_OP_GET_OBJECT_HANDLES, storage->storage_id, 0, MTP_OBJECT_HANDLE_ALL));
+    TEST_ASSERT_EQUAL_UINT32(mp3_count + 1, test_mtp_get_object_count(MTP_OP_GET_OBJECT_HANDLES, storage->storage_id, MTP_OBJ_FORMAT_MP3, MTP_OBJECT_HANDLE_ALL));
+    TEST_ASSERT_EQUAL_UINT32(all_count + 6, test_mtp_get_object_count(MTP_OP_GET_NUM_OBJECTS, storage->storage_id, 0, MTP_OBJECT_HANDLE_ALL));
+
+    TEST_ASSERT_EQUAL(ESP_OK, tinyusb_mtp_unregister_storage(storage));
+    TEST_ASSERT_EQUAL(ESP_OK, tinyusb_mtp_uninstall_driver());
+    TEST_ASSERT_EQUAL(ESP_OK, mtp_recursive_delete_path(tree_path));
 }
 
 TEST_CASE("MTP: Android direct edit backend updates object ranges", "[mtp][storage][ci]")
@@ -526,7 +780,8 @@ TEST_CASE("MTP: long ObjectInfo filename fits the endpoint buffer", "[mtp][stora
     }, &storage));
 
     uint32_t object_handle = 0;
-    TEST_ASSERT_EQUAL(ESP_OK, tinyusb_mtp_test_send_zero_size_object_info(storage, MTP_ROOT_PARENT, name, &object_handle));
+    TEST_ASSERT_EQUAL(ESP_OK, tinyusb_mtp_test_send_object_info(storage, TINYUSB_MTP_TEST_STORAGE_ID_REGISTERED, TINYUSB_MTP_TEST_STORAGE_ID_REGISTERED,
+                                                                MTP_ROOT_PARENT, name, 0, &object_handle));
     TEST_ASSERT_NOT_EQUAL(0, object_handle);
 
     struct stat st;
@@ -612,6 +867,49 @@ TEST_CASE("MTP: directory refresh removes stale cached objects", "[mtp][storage]
     TEST_ASSERT_EQUAL(ESP_OK, tinyusb_mtp_uninstall_driver());
     test_mtp_remove_path_if_exists(keep_path);
     test_mtp_remove_path_if_exists(dir_path);
+}
+
+TEST_CASE("MTP: full object cache reclaims missing entries", "[mtp][storage][ci]")
+{
+    test_mtp_storage_t *flash = &s_test_storages[0];
+    char path[128];
+    for (size_t i = 0; i < CONFIG_TINYUSB_MTP_MAX_OBJECTS; i++) {
+        char name[32];
+        snprintf(name, sizeof(name), "ci_cache_%03u.txt", (unsigned)i);
+        test_mtp_build_path(path, sizeof(path), flash->base_path, name);
+        test_mtp_remove_path_if_exists(path);
+    }
+    test_mtp_clean_storage_path(flash, "ci_cache_replacement.txt");
+
+    TEST_ASSERT_EQUAL(ESP_OK, tinyusb_mtp_install_driver(NULL));
+    tinyusb_mtp_storage_handle_t storage = NULL;
+    TEST_ASSERT_EQUAL(ESP_OK, tinyusb_mtp_register_storage(&(tinyusb_mtp_storage_config_t) {
+        .base_path = flash->base_path,
+        .display_name = flash->display_name,
+        .removable = flash->removable,
+    }, &storage));
+
+    for (size_t i = 0; i < CONFIG_TINYUSB_MTP_MAX_OBJECTS; i++) {
+        char name[32];
+        snprintf(name, sizeof(name), "ci_cache_%03u.txt", (unsigned)i);
+        test_mtp_build_path(path, sizeof(path), flash->base_path, name);
+        test_mtp_write_file(path, "cache");
+        uint32_t handle = 0;
+        TEST_ASSERT_EQUAL(ESP_OK, tinyusb_mtp_test_find_object(storage, path, &handle));
+    }
+    TEST_ASSERT_EQUAL_UINT32(CONFIG_TINYUSB_MTP_MAX_OBJECTS, test_mtp_cached_object_count(storage));
+
+    test_mtp_build_path(path, sizeof(path), flash->base_path, "ci_cache_000.txt");
+    TEST_ASSERT_EQUAL(0, unlink(path));
+    test_mtp_build_path(path, sizeof(path), flash->base_path, "ci_cache_replacement.txt");
+    test_mtp_write_file(path, "replacement");
+    uint32_t replacement_handle = 0;
+    TEST_ASSERT_EQUAL(ESP_OK, tinyusb_mtp_test_find_object(storage, path, &replacement_handle));
+    TEST_ASSERT_NOT_EQUAL(0, replacement_handle);
+
+    TEST_ASSERT_EQUAL(ESP_OK, tinyusb_mtp_unregister_storage(storage));
+    TEST_ASSERT_EQUAL(ESP_OK, tinyusb_mtp_uninstall_driver());
+    test_mtp_clean_storage(flash);
 }
 
 TEST_CASE("MTP: register three mounted FATFS paths", "[mtp][storage][ci]")
