+++ device/esp_tinyusb/test_apps/mtp_storage/sdkconfig.defaults
@@ -3,7 +3,7 @@ CONFIG_TINYUSB_MTP_ENABLED=y
 CONFIG_TINYUSB_MTP_TRACE_WRITES=n
 CONFIG_TINYUSB_MTP_EP_BUFSIZE=4096
 CONFIG_TINYUSB_MTP_MAX_STORAGES=3
-CONFIG_TINYUSB_MTP_MAX_OBJECTS=128
+CONFIG_TINYUSB_MTP_MAX_OBJECTS=32
 
 # Partitions configuration, used by spiflash FATFS storage.
 CONFIG_PARTITION_TABLE_CUSTOM=y
@@ -15,13 +15,11 @@ CONFIG_FATFS_LFN_HEAP=y
 CONFIG_FATFS_VOLUME_COUNT=4
 CONFIG_FATFS_CODEPAGE_936=y
 CONFIG_FATFS_API_ENCODING_UTF_8=y
+CONFIG_FREERTOS_HZ=1000
 
-# On chips with USB Serial JTAG, disable secondary console which does not make sense when using console component.
-CONFIG_ESP_CONSOLE_SECONDARY_NONE=y
 
 # Disable watchdogs, they'd get triggered during unity interactive menu.
 CONFIG_ESP_INT_WDT=n
-CONFIG_ESP_TASK_WDT=n
 CONFIG_ESP_MAIN_TASK_STACK_SIZE=8192
 
 # Run-time checks of Heap and Stack.
